graph-rag/requirements.txt
neo4j
faiss.cpu
google-generativeai
pypdf
numpy
python-dotenv
llama-cpp-python